requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
pymongo==4.13.2
python-dotenv==1.1.1
python-jose==3.5.0
email-validator==2.2.0
qrcode[pil]==8.2
